+++ src/main/java/com/ssafy/sgdc/competition/MatchingRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatchingRepo extends JpaRepository<Matching, Long> {
+public interface MatchingRepo extends JpaRepository<Matching, Integer> {
