+++ src/main/java/com/ssafy/sgdc/domain/entity/SsafyFriends.java
@@ -18,7 +18,7 @@ public class SsafyFriends {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
     @Column(name = "friend_id")
-    private long friendId;
+    private int friendId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
