+++ src/main/java/com/ssafy/sgdc/follow/dto/FollowingListResponseDto.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 public class FollowingListResponseDto {
 
     private String userNickname;
-    private Long userId;
+    private int userId;
     private String userImg;
 
 
-    public FollowingListResponseDto(Long userId, String userNickname, String userImg) {
+    public FollowingListResponseDto(int userId, String userNickname, String userImg) {
         this.userId = userId;
         this.userNickname = userNickname;
         this.userImg = userImg;
